Move 'Auth0Settings' into a named namespace
